refactor: 실제 로깅을 filter에서 하도록 변경
